+++ tests/unit/test_main.py
@@ -27,6 +27,7 @@ def test__parse_args__from():
         pattern_prefix=None,
         ignore_untracked=False,
         commit_length=None,
+        commit_prefix=None,
     )
     assert parse_args(["from", "git"]).vcs == "git"
     assert parse_args(["from", "git", "--tag-branch", "foo"]).tag_branch == "foo"
@@ -57,6 +58,7 @@ def test__parse_args__from():
     assert parse_args(["from", "any", "--pattern-prefix", "foo-"]).pattern_prefix == "foo-"
     assert parse_args(["from", "any", "--ignore-untracked"]).ignore_untracked is True
     assert parse_args(["from", "any", "--commit-length", "10"]).commit_length == 10
+    assert parse_args(["from", "any", "--commit-prefix", "x"]).commit_prefix == "x"
     assert parse_args(["from", "subversion", "--tag-dir", "foo"]).tag_dir == "foo"
 
     with pytest.raises(SystemExit):