+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## Unreleased
+
+* Added `commit_prefix` option to insert a prefix when serializing commit IDs.
+
 ## v1.24.1 (2025-05-09)
 
 * Fixed: Deprecation warning from `re.sub()` using Python 3.13+.